+++ Projects/DesignSystem/Sources/Component/Extension/UIView+sugar.swift
@@ -1,0 +1,23 @@
+//
+//  UIView+sugar.swift
+//  DesignSystem
+//
+//  Created by Eddy on 2022/07/24.
+//  Copyright © 2022 dvHuni. All rights reserved.
+//
+
+import UIKit
+
+extension UIView {
+    func createDottedLine(width: CGFloat, color: CGColor) {
+        let caShapeLayer = CAShapeLayer()
+        caShapeLayer.strokeColor = color
+        caShapeLayer.lineWidth = width
+        caShapeLayer.lineDashPattern = [2,4]
+        let cgPath = CGMutablePath()
+        let cgPoint = [CGPoint(x: self.bounds.minX, y: 0), CGPoint(x: 310, y: 0)]
+        cgPath.addLines(between: cgPoint)
+        caShapeLayer.path = cgPath
+        layer.addSublayer(caShapeLayer)
+    }
+}
