+++ Projects/DesignSystem/Sources/Component/ListCard/DDIPListCardView.swift
@@ -165,17 +165,3 @@ public class DDIPListCardView: UIView, AddViewsable {
         ])
     }
 }
-
-extension UIView {
-    func createDottedLine(width: CGFloat, color: CGColor) {
-        let caShapeLayer = CAShapeLayer()
-        caShapeLayer.strokeColor = color
-        caShapeLayer.lineWidth = width
-        caShapeLayer.lineDashPattern = [2,4]
-        let cgPath = CGMutablePath()
-        let cgPoint = [CGPoint(x: self.bounds.minX, y: 0), CGPoint(x: 310, y: 0)]
-        cgPath.addLines(between: cgPoint)
-        caShapeLayer.path = cgPath
-        layer.addSublayer(caShapeLayer)
-    }
-}
